feat: extract useSelectedNodeSettings hook from workbench
